+++ RELEASES.md
@@ -809,3 +809,13 @@ Based on:
 - [typescript v0.5.0-alpha.6] .
 ### Releases
 - [NPM v0.5.0-alpha.6] https://www.npmjs.com/package/cribl-control-plane/v/0.5.0-alpha.6 - .
+
+## 2026-01-19 01:57:44
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.686.0 (2.796.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.5.0-alpha.7] .
+### Releases
+- [NPM v0.5.0-alpha.7] https://www.npmjs.com/package/cribl-control-plane/v/0.5.0-alpha.7 - .